Struct oas_core::couch::ChangesStream

pub struct ChangesStream { /* private fields */ }

The stream for the _changes endpoint.

This is returned from [Database::changes].

Implementations

impl ChangesStream

pub fn new(db: CouchDB, last_seq: Option<String>) -> Self

Create a new changes stream.

pub fn with_params(
db: CouchDB,
last_seq: Option<String>,
params: HashMap<String, String>
) -> Self

Create a new changes stream with params.

pub fn set_last_seq(&mut self, last_seq: Option<String>)

Set the starting seq.

pub fn set_infinite(&mut self, infinite: bool)

Set infinite mode.

If set to true, the changes stream will wait and poll for changes. Otherwise, the stream will return all changes until now and then close.

pub fn set_selector(&mut self, selector: Value)

Set a selector to filter the changes stream.

See https://docs.couchdb.org/en/stable/api/database/changes.html#selector for details.

pub fn set_param(&mut self, key: impl ToString, value: impl ToString)

pub fn last_seq(&self) -> &Option<String>

Get the last retrieved seq.

pub fn infinite(&self) -> bool

Whether this stream is running in infinite mode.

pub fn batched_untyped_records(
self,
opts: &BatchOpts
) -> impl Stream<Item = UntypedRecordBatch>